+++ hutool-core/src/main/java/cn/hutool/core/thread/threadlocal/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 
+ * ThreadLocal相关封装
+ * @author looly
+ *
+ */
+package cn.hutool.core.thread.threadlocal;